requirements.txt
streamlit
fer
opencv-python
tensorflow
Pillow
protobuf==3.20.1
